crud resolved

+++ src/main/java/app/dao/UserDao.java
@@ -9,9 +9,9 @@ public interface UserDao {
     void addUser(User user);
     User getUserByName(String name);
     void updateUser(User user);
-    void deleteUser(int userId);
+    void deleteUser(long userId);
     List<User> getAllUsers(int offSet, int limit);
-    User getUserById(int userId);
+    User getUserById(long userId);
     List<User> getLastUsers(int limit);
 
 }
